requirements.txt
streamlit>=1.40.0
yfinance>=0.2.48
pandas>=2.2.0
plotly>=5.24.0
requests>=2.32.0
google-genai>=1.0.0
